# Fix parsing of crypto tick data with sub-millisecond timestamps