--- a/Tests/Common/Data/Market/TickTests.cs
+++ b/Tests/Common/Data/Market/TickTests.cs
@@ -1,11 +1,11 @@
 ﻿/*
  * QUANTCONNECT.COM - Democratizing Finance, Empowering Individuals.
  * Lean Algorithmic Trading Engine v2.0. Copyright 2014 QuantConnect Corporation.
- * 
- * Licensed under the Apache License, Version 2.0 (the "License"); 
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -39,6 +39,26 @@ namespace QuantConnect.Tests.Common.Data.Market
             Assert.AreEqual("P", tick.Exchange);
             Assert.AreEqual("T", tick.SaleCondition);
             Assert.AreEqual(false, tick.Suspicious);
+        }
+
+        [Test]
+        public void ConstructsFromLineWithDecimalTimestamp()
+        {
+            const string line = "18000677.3,3669.12,0.0040077,3669.13,3.40618718";
+
+            var config = new SubscriptionDataConfig(
+                typeof(Tick), Symbols.BTCUSD, Resolution.Tick, TimeZones.Utc, TimeZones.Utc,
+                false, false, false, false, TickType.Quote);
+            var baseDate = new DateTime(2019, 1, 15);
+
+            var tick = new Tick(config, line, baseDate);
+
+            var ms = (tick.Time - baseDate).TotalMilliseconds;
+            Assert.AreEqual(18000677, ms);
+            Assert.AreEqual(3669.12, tick.BidPrice);
+            Assert.AreEqual(0.0040077, tick.BidSize);
+            Assert.AreEqual(3669.13, tick.AskPrice);
+            Assert.AreEqual(3.40618718, tick.AskSize);
         }
 
         [Test]