#### Description
- The `Tick` constructor has been fixed to include parsing timestamps with decimals
- The extension methods `ToInt32` and `ToInt64` have been updated to support input strings with decimals

#### Related Issue
Closes #2844

#### Motivation and Context
Crypto Tick live history requests were returning empty results.

#### Requires Documentation Change
No.

#### How Has This Been Tested?
New unit tests + history calls with live algorithms.

#### Types of changes
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description` or `feature-<issue#>-<description>`