@jaredbroad 2019-01-21
Pending test on the tick regression algorithm speed before and after this change.