@StefanoRaggi 2019-01-21
Benchmark algorithm:
``` C#
    public class BasicTemplateAlgorithm : QCAlgorithm
    {
        public override void Initialize()
        {
            SetStartDate(2018, 1, 1);
            SetEndDate(2018, 12, 31);
            SetCash(100000);

            AddCrypto("BTCUSD", Resolution.Tick);
        }
    }
```

Before - Master (v5063):

```
Algorithm Id:(bb373a62570b7bd3c7b8ec33068189eb) completed in 171.51 seconds at 146k data points per second. Processing total of 24,955,237 data points.
Algorithm Id:(2a87a8fc2173f6337a2b410c614097a7) completed in 159.69 seconds at 156k data points per second. Processing total of 24,955,237 data points.
Algorithm Id:(b4fdc8e248446e9215ad16242c65d4a7) completed in 169.18 seconds at 148k data points per second. Processing total of 24,955,237 data points.
```

After - Branch (v5068):

```
Algorithm Id:(43d5ade0456ca11b294a8b13c2d49dac) completed in 164.18 seconds at 152k data points per second. Processing total of 24,955,237 data points.
Algorithm Id:(1a28308b318f8808f4d143e8b734e2e8) completed in 163.05 seconds at 153k data points per second. Processing total of 24,955,237 data points.
Algorithm Id:(51c3131bbed6cdd9278ae4304b67c207) completed in 164.41 seconds at 152k data points per second. Processing total of 24,955,237 data points.
```